backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.19.0
pytest==7.3.1
flake8==6.0.0
gunicorn==20.1.0
Werkzeug==2.2.3